feat(ai-defect): split hallucination rules
Move strict hallucination checks into rules/ai_defect and route SKY-L012, SKY-L023, SKY-D222, SKY-D224, and SKY-D225 into the ai_defects output bucket. Keep old quality/danger import paths as compatibility wrappers.

Pre-commit note: local staged gate flags pre-existing complexity in mechanically moved scanner files. This commit preserves the moved implementations exactly; focused tests passed (607 passed) and docs build passed.
