+++ skylos/rules/catalog.py
@@ -58,14 +58,14 @@ _RULES = (
     RuleCatalogEntry("SKY-L009", "Debug leftover", "quality"),
     RuleCatalogEntry("SKY-L010", "Security TODO marker", "quality"),
     RuleCatalogEntry("SKY-L011", "Disabled security control", "quality"),
-    RuleCatalogEntry("SKY-L012", "Phantom function call", "quality"),
+    RuleCatalogEntry("SKY-L012", "Phantom function call", "ai_defect"),
     RuleCatalogEntry("SKY-L013", "Insecure randomness", "quality"),
     RuleCatalogEntry("SKY-L014", "Hardcoded credential", "quality"),
     RuleCatalogEntry("SKY-L016", "Undefined config", "quality"),
     RuleCatalogEntry("SKY-L017", "Error information disclosure", "quality"),
     RuleCatalogEntry("SKY-L020", "Overly broad file permissions", "quality"),
     RuleCatalogEntry("SKY-L021", "Security regression", "quality"),
-    RuleCatalogEntry("SKY-L023", "Phantom decorator", "quality"),
+    RuleCatalogEntry("SKY-L023", "Phantom decorator", "ai_defect"),
     RuleCatalogEntry("SKY-L024", "Stale mock", "quality"),
     RuleCatalogEntry("SKY-L026", "Unfinished generated code", "quality"),
     RuleCatalogEntry("SKY-L027", "Duplicate string literal", "quality"),
@@ -105,10 +105,10 @@ _RULES = (
     RuleCatalogEntry("SKY-D215", "Path traversal", "security", "HIGH", aliases=("lfi", "file traversal")),
     RuleCatalogEntry("SKY-D216", "Server-side request forgery", "security", "CRITICAL", aliases=("ssrf",),),
     RuleCatalogEntry("SKY-D217", "Raw SQL execution", "security", "CRITICAL"),
-    RuleCatalogEntry("SKY-D222", "Dependency hallucination", "security", "CRITICAL"),
+    RuleCatalogEntry("SKY-D222", "Dependency hallucination", "ai_defect", "CRITICAL"),
     RuleCatalogEntry("SKY-D223", "Undeclared third-party dependency", "security", "MEDIUM"),
-    RuleCatalogEntry("SKY-D224", "API signature hallucination", "security", "HIGH"),
-    RuleCatalogEntry("SKY-D225", "Dependency version hallucination", "security", "HIGH"),
+    RuleCatalogEntry("SKY-D224", "API signature hallucination", "ai_defect", "HIGH"),
+    RuleCatalogEntry("SKY-D225", "Dependency version hallucination", "ai_defect", "HIGH"),
     RuleCatalogEntry("SKY-D226", "Cross-site scripting: untrusted content marked safe", "security", "CRITICAL",aliases=("xss", "cross site scripting", "cross-site scripting"),),
     RuleCatalogEntry("SKY-D227", "Cross-site scripting: unsafe inline template", "security", "HIGH", aliases=("xss", "cross site scripting", "cross-site scripting"),),
     RuleCatalogEntry("SKY-D228", "Cross-site scripting: HTML built from user input", "security", "HIGH", aliases=("xss", "cross site scripting", "cross-site scripting"),),
