+++ skylos/ui/terminal_report.py
@@ -11,6 +11,7 @@ from skylos.api._snippets import _resolve_snippet_path
 
 
 CATEGORY_SPECS = (
+    ("ai_defects", "AI Defect", "AI defect"),
     ("danger", "Security", "security issue"),
     ("secrets", "Secret", "secret detected"),
     ("quality", "Quality", "quality issue"),
